[FEAT] 대학 리스트 분리, UserInfoScreen 검색 기능 추가
