+++ app/src/main/java/com/example/roomfit/presentation/UserEditScreen.kt
@@ -27,6 +27,7 @@ import androidx.navigation.compose.rememberNavController
 import coil.compose.rememberAsyncImagePainter
 import com.example.roomfit.R
 import com.example.roomfit.presentation.components.LoginButton
+import com.example.roomfit.presentation.login.UnivList
 import com.example.roomfit.ui.theme.*
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -52,6 +53,7 @@ fun UserEditScreen(
     var lifestyleState by remember { mutableStateOf(lifestyle) }
     var smokingState by remember { mutableStateOf(smoking) }
     var profileImageUri by remember { mutableStateOf<Uri?>(null) }
+    var showUniversityDialog by remember { mutableStateOf(false) }
 
     val houseTypes = listOf("아파트", "빌라", "원룸")
     val residentNumbers = listOf("2명", "3명", "4명 이상")
@@ -147,14 +149,27 @@ fun UserEditScreen(
                 onValueChange = { schoolState = it },
                 label = "나의 대학 찾기",
                 trailingIcon = {
-                    IconButton(onClick = { /* 아이콘 클릭 시 동작 */ }) {
-                        Icon(
+                    IconButton(
+                        onClick = { showUniversityDialog = true },
+                        modifier = Modifier.padding(15.dp)
+                    ) {
+                        Image(
                             painter = painterResource(id = R.drawable.school_search),
                             contentDescription = "Search School"
                         )
                     }
+
+                    if (showUniversityDialog) {
+                        UnivList(
+                            onDismiss = { showUniversityDialog = false },
+                            onUniversitySelected = { selectedUniversity ->
+                                schoolState = selectedUniversity
+                                showUniversityDialog = false
+                            }
+                        )
+                    }
                 }
-            )
+            ) // 여기 닫는 괄호 추가
 
             Spacer(modifier = Modifier.height(50.dp))
